clickhouse query builder
